TraceArrays

Documentation for TraceArrays.

TraceArrays.AbstractFourierTraceArray

TraceArrays.AbstractTraceArray

TraceArrays.DASArray

TraceArrays.DASArray

TraceArrays.DASArray

TraceArrays.DASArray

TraceArrays.DASArray

TraceArrays.FourierDASArray

TraceArrays.FourierDASArray

TraceArrays.TraceArray

TraceArrays.TraceArray

TraceArrays.TraceArray

AbstractFFTs.fft

Base.append!

Base.empty

Base.reverse!

Base.vcat

Seis.integrate!

Seis.normalise!

Seis.remove_trend!

Seis.taper!

TraceArrays._average_channel_positions

TraceArrays._average_position

TraceArrays._check_abstracttracearray_consistency

TraceArrays._check_trace_consistency

TraceArrays._differentiate

TraceArrays._integrate!

TraceArrays._integrate_rectangle!

TraceArrays._integrate_trapezium

TraceArrays._matrix_from_traces

TraceArrays._normalise_all!

TraceArrays._normalise_columns!

TraceArrays.cut_distance

TraceArrays.cut_distance!

TraceArrays.decimate_distance

TraceArrays.decimate_distance!

TraceArrays.differentiate_distance

TraceArrays.differentiate_distance!

TraceArrays.distances

TraceArrays.integrate_distance

TraceArrays.nwavenumbers

TraceArrays.read_febus

TraceArrays.remove_median_trace

TraceArrays.remove_median_trace!

TraceArrays.wavenumbers

TraceArrays.AbstractFourierTraceArray — Type

AbstractFourierTraceArray <: AbstractFourierTrace

The abstract AbstractFourierTraceArray type is the supertype for trace array types which are in the frequency (Fourier) domain.

TraceArrays.AbstractTraceArray — Type

AbstractTraceArray <: Seis.AbstractTrace

The abstract AbstractTraceArray type is the supertype for types which correspond to sets of recordings of seismic data which are, strictly interpreted, 'arrays'. This means that all sensors have a common time base and are of the same kind (i.e., have the same response).

The data underlying AbstractTraceArrays is usually stored as a matrix with each column containing a continuous time series of data, and each row representing the same time point at each array element (although the exact implementation of AbstractTraceArray subtype need not be exactly like this).

Indexing (getindex)

Indexing into a AbstractTraceArray with a single index gives you a single Trace back; single-dimension slicing gives you a vector of Traces.

Note

In a future version of TraceArrays (or when it is merged into Seis.jl) slicing will instead return a new AbstractTraceArray.

Required fields

Subtypes of AbstractTraceArray should have at least the following fields:

b: Time in s relative to the origin time (if any) set in evt.time

delta: Sampling interval in s

evt::Event: Event relating to these data. .evt.time is the time to which .b is relative.

sta::AbstractVector{<:Station}: Set of stations, one for each channel

data::AbstractMatrix: Data, where each column holds a single channel

picks::Seis.SeisDict{Union{Int,Symbol}, Seis.Pick{T}}: Set of picks, where any one pick represents a single time across all channels.

meta::Seis.SeisDict{Symbol,Any}: Metadata for all channels.

Required methods

empty(::T) where {T<:AbstractTraceArray}: Constructor for a new empty AbstractTraceArray where all fields are deepcopyed apart from .data. This is intended to be used to implement copying functions (e.g., Seis.decimate) which do not need to first copy the whole .data field.

Fallback methods

The following are defined for all AbstractTraceArrays, based on their containing the above fields:

Seis.trace(::AbstractTraceArray)::AbstractMatrix: The data matrix

Seis.nsamples(::AbstractTraceArray)::Int: Number of samples

Seis.times(::AbstractTraceArray): Range giving times of each sample

Base.eltype(::AbstractTraceArray)::DataType: Element type of data matrix

Base.length(::AbstractTraceArray)::Int: Number of channels

Base.getindex(::AbstractTraceArray, i::Int)::AbstractTrace

Base.getindex(::AbstractTraceArray, indices)::AbstractVector{<:AbstractTrace}

Base.getindex(::AbstractTraceArray, ::Colon)::AbstractVector{<:AbstractTrace}

Base.firstindex(::AbstractTraceArray)

Base.lastindex(::AbstractTraceArray)

(Base.empty(::T) where {T<:AbstractTraceArray})::T

_check_abstracttracearray_consistency(::AbstractTraceArray, ::Vararg{AbstractTraceArray}): Whether two or more arrays can be append!ed or vcatted together.

Generated Base functions

== and hash are defined recursively on the fields of subtypes of AbstractTraceArray. These can be overloaded if needed.

TraceArrays.DASArray — Type

DASArray{T,M,P<:Seis.Position} <: AbstractTraceArray

AbstractTraceArray holding DAS data.

This type differs from TraceArray in that it records the information on gauge (channel) spacing and distance, which is useful for some things.

DASArray(; kwargs...)

Return a new, empty object for DAS data. Individual fields may be defined using keyword arguments.

Note

The keyword argument constructor for DASArrays is not a stable part of the API and may change.

Keyword arguments

data: AbstractMatrix of data, where columns contain continuous evenly-sampled recordings at each distance, and each column is recorded a constant distance from the last. These columns therefore correspond to the gauges used in DAS recordings. The number of channels (columns) is used to create the correct number of stations in the .sta field.

starting_distace: Distance from the interrogator of the first channel (m)

distance_spacing: Spacing between channels (m)

b: Start time (relative to .evt.time) of the first sample (s)

delta: Sampling interval (s)

TraceArrays.DASArray — Method

DASArray(array_of_traces::AbstractArray{<:AbstractTrace}, starting_distance, distance_spacing)

Construct a DASArray from an array of single-channel traces, providing the distance to the first channel in m as starting_distance, and the channel spacing in m as distance_spacing.

TraceArrays.DASArray — Method

DASArray(t::FebusTools.FebusData)

Construct a DASArray from data read from a Febus A1 interrogator.

Instrument metadata is placed in the trace's .meta.febus field. Note that some metadata items (such as :Extent) are not modified from the file read, and so may not be in sync with the data returned.

TraceArrays.DASArray — Method

DASArray(::AbstractTraceArray, starting_distance, distance_spacing) -> ::DASArray

Construct a DASArray from another kind of AbstractTraceArray (like a TraceArray). Supply the distance of the first channel in m as starting_distance and the channel spacing in distance_spacing, also in m.

TraceArrays.DASArray — Method

DASArray(t::DASArray{T,M,P}; kwargs...) -> t′

Create a new DASArray from an existing one, but where the trace data is replaced by data. The data matrix type (M parameter in the DASArray type) is replaced by the type of data.

If the number of columns in data is not the same as the number of channels in t, then the vector of stations in t′.sta is adjusted to be the same length as data. If there are more channels than before, the t′.sta vector will contain #undef elements where the additional channels are present. Accessing these elements before properly setting them will throw an UndefRefError.

TraceArrays.FourierDASArray — Type

FourierDASArray <: AbstractFourierTraceArray

The Fourier transform of a DASArray.

Because DASArrays by definition are equally sampled in space along the fibre, FourierDASArrays represent the 2D Fourier transform of the data sampled. FourierDASArrays contain all the information needed on the original number of samples (in time) and channels (in space). The fields b, delta (frequency spacing in Hz), starting_distance distance_spacing, sta, evt, picks and meta remain part of the public API of the type.

The usual way to create FourierDASArrays is by calling fft on a DASArray. To convert back to a DASArray, call ifft

Note

FourierDASArrays share many fields with the DASArray they came from, including the station and event, plus picks and metadata. These fields are not copied across, and are instead references to the same Station, Event, and so on. Therefore, any changes to the FourierDASArray will be reflected in the corresponding DASArray. To avoid this, do f = fft(deepcopy(t)).

See also: DASArray, fft, ifft.

TraceArrays.FourierDASArray — Method

FourierDASArray(; b, delta, data, nsamples=(2*length(data) - 1), starting_distance, distance_spacing, evt=Event(), sta=Station(), picks=nothing, meta=nothing)

Create a FourierDASArray directly from a set of Fourier coefficients. Users will usually construct a FourierDASArray by calling fft on a Trace instead of using this constructor.

Keyword arguments

b: The starting time in s of the original recording this frequency domain trace represents.

delta: The original sampling interval in s of the equivalent time domain trace.

data: An AbstractMatrix{<:Complex} containing the set of Fourier coefficients for this frequency domain trace. Note that this is a 'one-sided' set, where the first index of the first dimension corresponds to 0 Hz and the final index is the Nyquist frequency. This is because DASArrays represent real quantities. In the second dimension, the first index corresponds to the most negative wavenumber, the central index a wavenumber of 0, and the final index the largest wavenumber, where the most negative and largest wavenumbers have the same absolute value.

nsamples: The number of time domain samples in the origin time domain trace which this frequency domain trace represents. This allows the FourierDASArray to be converted back to the original DASArray with no loss of the number of points.

starting_distance: Distance along cable of the first channel (m).

distance_spacing: Spacing between channels along the cable (m).

evt: An Event which defines the source and origin time for the data. Normally this should be taken from the Trace being used to construct the FourierDASArray.

sta: A Vector{<:Station} which defines the recording stations for the data. Normally this should be taken from the DASArray being used to construct the FourierDASArray.

TraceArrays.TraceArray — Type

TraceArray{T,M,P} <: AbstractTraceArray

TraceArrays are sets of traces which are recorded on a strictly defined array of seismic recorders. This strict definition requires all channels to have the same response and for all samples to be recorded at the same time across array elements (meaning a uniform sampling interval as well). For a single TraceArray, no channels can be partially present, meaning all start and end on the same sample (although a TraceArray can be constructed from channels where gaps have been filled).

TraceArrays.TraceArray — Method

TraceArray(array_of_traces) -> ::TraceArray

Construct a TraceArray from an array of AbstractTraces.

Example

julia> using Seis

julia> TraceArray(filter(is_vertical, sample_data(:regional)))

TraceArrays.TraceArray — Method

TraceArray(; kwargs...)

Return a new, empty object for trace array data. Individual fields may be defined using keyword arguments.

Note

The keyword argument constructor for TraceArrays is not a stable part of the API and may change.

Keyword arguments

b: Start time (relative to .evt.time) of the first sample (s)

delta: Sampling interval (s)

data: AbstractMatrix of data, where columns contain continuous evenly-sampled recordings at each distance, and each column is recorded a constant distance from the last. These columns therefore correspond to the gauges used in DAS recordings. The number of channels (columns) is used to create the correct number of stations in the .sta field.

AbstractFFTs.fft — Method

Seis.fft(t::DASArray)

Convert the DAS data t into its equivalent frequency domain data f by performing a Fourier transform. The object returned is a FourierDASArray.

Example

julia> v_app = 5 # Apparent velocity m/s;

julia> freq = 0.5 # Wave frequency, Hz;

julia> delta = 0.05 # Sampling interval;

julia> distance_spacing = 0.05 # Channel spacing, m;

julia> t = DASArray(;
           b=0, delta, starting_distance=0, distance_spacing,
           data=sin.(2π*freq.*((0:201).*delta .- (0:201)'.*distance_spacing./v_app))
       ); # Sine wave of frequency 1 Hz with apparent velocity of 5 m/s;

julia> T = fft(t);

julia> indmax = argmax(abs.(trace(T))) # Get index of maximum power
CartesianIndex(6, 101)

julia> f = frequencies(T)[indmax[1]] # Maximum frequency in Hz is ~1 as expected
0.495049504950495

julia> wavenumbers(T) # Wavenumbers corresponding to second dimension of `trace(f)`
-9.999999999999998:0.099009900990099:9.900990099009901

julia> λ⁻¹ = -wavenumbers(T)[indmax[2]] # Wavenumber of wave in m⁻¹ is -1/-5 ≈ 0.2 as expected
0.09900990099009732

julia> f/λ⁻¹ # Wave velocity is frequency over wavenumber; ~5 m/s as expected
5.000000000000084

See also: ifft.

Base.append! — Method

append!(a::AbstractTraceArray, rest::AbstractTraceArray...; merge_meta=true) -> a

Append the traces in rest to a, requiring that they all have consistent start times, sampling rates and trace lengths.

Metadata are merged together by default (unless merge_meta is false). If any entries of the same name occur in both, then those in a are used in preference; otherwise later entries in the argument list are kept.

Base.empty — Method

empty(t::T) where {T<:AbstractTraceArray} -> t′::T

Create a copy of an AbstractTraceArray t where all fields apart from the underlying trace data are copied to a new instance of T using Base.deepcopy.

This function is not exported and is meant to be used by library code to efficiently implement in- and out-of-place processing functions.

t′.data is set to be a 0×nchannels, matrix. It is the caller's responsibility to ensure that the matrix is set to be consistent with the other internal fields (chiefly .sta) before t′ is used further.

Base.reverse! — Method

reverse!(t::AbstractTraceArray) -> t

Reverse the order of the channels in t.

Base.vcat — Method

Base.vcat(a::AbstractTraceArray, rest::AbstractTraceArray...; merge_meta=true)

Append the traces in rest to a, requiring that they both have consistent start times, sampling rates and trace lengths, returning copies.

Metadata are merged together by default (unless merge_meta is false). If any entries of the same name occur, then those in a are used in preference; otherwise later entries in the argument list are kept.

Seis.integrate! — Function

Seis.integrate!(t::AbstractTraceArray, method=:trapezium) -> t

Integrate each channel in t, replacing the underlying data with a copy.

Seis.normalise! — Function

normalise!(t::AbstractTraceArray, val=1; all=false) -> t

Normalise all channels in t to have maximum absolute value val.

By default, this is done on a channel-by-channel basis. To normalise all channels simultaneously, preserving the inter-channel relative amplitudes, use all=true. In this case, the single maximum absolute amplitude across all the data will be normalised to val instead.

Seis.remove_trend! — Method

Seis.remove_trend!(t::AbstractTraceArray)

Remove the linear trend from all channels of t, where the trend is found by fitting a straight line to the linear stack of all channels.

Seis.taper! — Function

Seis.taper!(t::AbstractTraceArray, width=0.05; form=:hanning)

Taper all channels in t with a taper of width, which is a fraction of the whole trace between 0 and 0.5.

TraceArrays._average_channel_positions — Method

_average_channel_positions(stas::AbstractArray{<:Seis.Station}, points) -> pos::Vector{<:Seis.Position}

Return a vector of positions which are the average channel positions for a points-point finite difference differentiation in space. pos will be shorter than stas due to the differentiation.

TraceArrays._average_position — Method

_average_position(a::Seis.Position, b::Seis.Position) -> pos

Return the mean position of a and b obtained by simple averaging of each coordinate separately. This is fine for Seis.Cartesian, but may be bad for Seis.Geographic if a and b are far apart. A warning is thrown in that case.

TraceArrays._check_abstracttracearray_consistency — Method

_check_abstracttracearray_consistency(ts)

Function which should throw if any AbstractTraceArrays cannot be sensible merged together with vcat or append.

TraceArrays._check_trace_consistency — Method

_check_trace_consistency(ts::NTuple{N,AbstractTraceArray} where N)

Throw an error if the traces in ts cannot be sensibly joined together into a new AbstractTraceArray with vcat or append!.

New subtypes of AbstractTraceArray may implement methods for this function which impose additional checks on whether all the items in ts are compatible with each other.

TraceArrays._differentiate — Method

_differentiate(data, points, b, delta) -> data′, b

Differentiate the matrix data along its columns, assuming each column is an evenly-sampled series with spacing delta starting at b seconds. Return the differentiated matrix data′ and new start time b′.

TraceArrays._integrate! — Method

_integrate!(U, u, spacing_by_2, n) -> U

Integrate the values of u using an n-point trapezium rule, putting the output into U. U must be the correct size, i.e., it must have n - 1 points fewer than u. spacing_by_2 is half the even spacing between points in u.

Example

julia> u = 1:7;

julia> U = similar(u, 5);

julia> TraceArrays._integrate!(U, u, 10/2, 3)
5-element Vector{Float64}:
  40.0
  60.0
  80.0
 100.0
 120.0

TraceArrays._integrate_rectangle! — Method

_integrate_rectangle!(m::AbstractMatrix, delta) -> m

For an input matrix m whose columns represent some function evaluated at evenly-spaced points delta apart, replace the columns in m with the integrated values.

TraceArrays._integrate_trapezium — Method

_integrate_trapezium(m::AbstractMatrix, delta) -> ∫m

For an input matrix m whose columns represent some function evaluated at evenly-spaced points delta apart, return an updated matrix ∫m whose columns contain the integrated columns of m. Note therefore that ∫m has one less row than m.

TraceArrays._matrix_from_traces — Method

_matrix_from_traces(array_of_traces) -> ::Matrix, T, M, P

Assuming that all the AbstractTraces in array_of_traces have the same number of samples, sampling rate and start time, construct a matrix of the data where each column contains the samples from the corresponding trace.

Also return T, M and P, the type parameters for the TraceArray which can be formed with the matrix.

TraceArrays._normalise_all! — Method

_normalise_all!(data, val)

Normalise all columns in data such that the maximum absolute value in the whole matrix is val.

TraceArrays._normalise_columns! — Method

_normalise_columns!(data::AbstractMatrix, val)

Normalise all columns in data to have the maximum absolute value val.

TraceArrays.cut_distance! — Method

cut_distance!(t::DASArray, x1, x2; warn=true, allowempty=false) -> t

Cut out channels in the DASArray t which lie between distances x1 m and x2 m. Only channels which are at exactly x1 m or more, and exactly x2 m or less are retained.

If allowempty is true, then an empty DASArray can be returned when x1 and x2 are either both below or above the cable distance range. Otherwise, an error is thrown.

By default, warnings are printed if x1 or x2 lie outside the existing range of distances for t and the start or end distance is used instead to perform the cut. Use warn=false to turn these warnings off.

See also: cut_distance.

TraceArrays.cut_distance — Method

cut_distance(t::DASArray, x1, x2; warn=true, allowempty=false) -> t′

Out-of-place version of cut_distance!.

TraceArrays.decimate_distance! — Method

decimate_distance!(t::DASArray, n::Integer) -> t

Naively decimate the channels in t, taking every nth channel and removing the rest. The data field of t is a copy of the data originally used, and the original data array is left untouched.

TraceArrays.decimate_distance — Method

decimate_distance(t::DASArray, n::Integer) -> t

Copying version of decimate_distance.

See also: decimate_distance.

TraceArrays.differentiate_distance! — Method

differentiate_distance!(t::DASArray; points::Integer=2) -> t

In-place version of differentiate_distance.

TraceArrays.differentiate_distance — Method

differentiate_distance(t::DASArray; points::Integer=2) -> t′

Differentiate the data in t with respect to distance along the cable using points-point finite differencing.

If data are in units of strain, they are converted to be proportional to units of length-squared. No unit conversion is made, so data which are in nanostrain convert to nm², nanostrain per second go to nm²/s, and so on.

The number of channels is reduced in the operation, and each channels coordinates are averaged in the new trace.

Available algorithms

(Pseudocode descriptions below drop the time index for clarity.)

points == 2: Two-point. t′.data[i] = (t.data[i+1] - t.data[i])/t.distance_spacing. Non-central difference, so t.starting_distance is increased by half t.distance_spacing. The number of channels is reduced by 1.

points == 3: Three-point. t′.data[i] = (t.data[i+1] - t.data[i-1])/(2 * t.distance_spacing). Central difference. t.starting_distance is increased by t.distance_spacing; the number of channels is reduced by 2.

points == 5: Five-point. `t′.data[i] = (2/3)*(t.data[i+1] - t.data[i-1])/t.distance_spacing

(1/12)*(t.data[i+2] - t.data[i-2])/t.distance_spacing`,

except for the first and last points, which use a three-point central difference meaning only two points fewer are retained as for points == 3. Central difference. t.starting_distance is increased by t.distance_spacing; number of channels reduced by 2.

TraceArrays.distances — Method

distances(da::DASArray) -> dists

Return a range giving the distances in m along the fibre of each channel in da.

TraceArrays.integrate_distance — Function

integrate_distance(t::DASArray, n=3) -> t′

Integrate the data in t with respect to distance along the cable. If data are in units of strain, therefore, they are converted to units of displacement (and similarly for strain rate to velocity). No unit conversion is made, so data which are in nanostrain convert to nm, and likewise nanostrain/s becomes nm/s.

n determines how many points are used in the integration and must be an odd number. This determines the number of points used for the trapezoidal integration.

The number of channels is reduced by n - 1. For example, when n is 3, t′ has 2 fewer channels than before, and the first and last channels in t are removed from t′.

Example

julia> t = DASArray(; data=[0. 1 0 -1; 1 2 -2 1], b=0, delta=1, starting_distance=0, distance_spacing=0.5);

julia> integrate_distance(t)
2×2 Matrix{Float64}:
 0.5    0.0
 0.75  -0.25

TraceArrays.nwavenumbers — Method

nwavenumbers(f::FourierDASArray) -> n

Return the number of wavenumbers in the frequency-domain DAS array f which are distinct only in terms of magnitude. For a trace array with original length N, this is N÷2 + 1.

TraceArrays.read_febus — Method

read_febus(file; kwargs...) -> ::DASArray{Float64, Matrix{Float32}, Seis.Geographic{Float64}}

Read DAS data from file in the Febus HDF5 format.

Keyword arguments

xlim=(-Inf, Inf): Read only a range of distances (m)

xdecimate=1: Decimate the number of channels read in.

tlim=(-146138511-01-01T00:00:00, 146138512-12-31T23:59:59): Read data between two DateTimes

blocks=(1,nblocks): Read in only data blocks within the given range. (Data blocks are typically 1 s long, but can be any length.)

TraceArrays.remove_median_trace! — Method

remove_median_trace!(t::AbstractTraceArray) -> t

Remove the median trace from all channels in t.

The operation finds the median value at each time sample across all channels and subtracts this from all channels, modifying the original trace.

See also remove_median_trace.

TraceArrays.remove_median_trace — Method

remove_median_trace(t::AbstractTraceArray) -> t′

Out-of-place version of remove_median_trace!.

TraceArrays.wavenumbers — Method

wavenumbers(f::FourierDASArray)

Return the wavenumbers (spatial frequencies) of the Fourier-domain DAS array f for each point of the second dimension of the underlying trace data obtained with trace(f). The units are m⁻¹, i.e., there is no factor of 2π.